feat: Refactor database entities and remove SyllabusEntity
This commit refactors the database entities to use `Long` as the primary key type instead of `Int`. It also removes the `SyllabusEntity` as it's not needed. The changes include:

-   Modify `SchoolEntity`: Change `schoolId` type to Long?.
-   Modify `LearningMaterialEntity`: Change `learningMaterialId` type to Long? and `lessonTopicId` to Long.
-   Modify `StudentEntity`: Change `studentId` type to Long? and `classId` to Long.
-   Modify `ExamPaperEntity`: Change `examPaperId` type to Long?, `courseId`, and `creatorStaffId` to Long, and `examScheduleId` to Long?.
-   Modify `LessonTopicEntity`: Change `topicId` type to Long? and `courseId` to Long.
-   Modify `ExamQuestionEntity`: Change `questionId` type to Long?, `examPaperId` to Long, `marks` to Long, and `lessonTopicId` to Long?.
-   Modify `LearningObjectiveEntity`: Change `learningObjectiveId` type to Long? and `lessonTopicId` to Long.
-   Modify `ExamScheduleEntity`: Change `examScheduleId` type to Long?, `classId` to Long, and `examPaperId` to Long?.
-   Modify `ChoiceOptionEntity`: Change `optionId` type to Long? and `examQuestionId` to Long.
-   Modify `AcademicStaffEntity`: Change `staffId` type to Long?.
-   Modify `StudentAnswerSheetEntity`: Change `answerSheetId` type to Long?, `studentId`, and `examScheduleId` to Long.
-   Modify `ClassEntity`: Change `classId` type to Long?, `gradeLevelId` to Long, and `teacherStaffId` to Long?.
-   Modify `StudentAnswerEntity`: Change `studentAnswerId` type to Long?, `answerSheetId`, `examQuestionId` to Long, and `choiceOptionId` to Long?. change `marksObtained` to long.
-   Modify `CourseGradeEntity`: Change `studentId` and `courseId` type to Long.
-   Modify `ClassAttendanceEntity`: Change `classAttendanceId` to Long?, `studentId`, and `classId` to Long.
-   Modify `ExamAttendanceEntity`: Change `examAttendanceId` type to Long?, `studentId`, and `examScheduleId` to Long.
-   Modify `CourseEntity`: Change `courseId` type to Long? and `gradeLevelId` to Long.
-   Modify `GradeLevelEntity`: Change `gradeLevelId` type to Long? and `levelNumber` to Long.
   -   Remove `SyllabusEntity` file.
-  - add `TeacherCourseQualificationEntity` entity.
-  - modify  `ExamInstructionEntity` : change  `ExaminationEntity` to  `ExamPaperEntity`
-  - change  `questionType` : from string to `QuestionType`
+  - add `TeacherCourseQualificationEntity`
+  - modify  `ExamInstructionEntity` : change foreign key `ExaminationEntity` to `ExamPaperEntity`
+  - change  `questionType` : from string to `QuestionType` enum
- -  use  `kotlinx.datetime.LocalDate` instead  `java.time.LocalDate`
- -  use  `kotlinx.datetime.LocalDateTime` instead  `java.time.LocalDateTime`
- -  use  `kotlinx.datetime.LocalTime` instead  `java.time.LocalTime`
+ - use  `kotlinx.datetime.LocalDate` instead  `java.time.LocalDate`
+ - use  `kotlinx.datetime.LocalDateTime` instead  `java.time.LocalDateTime`
+ - use  `kotlinx.datetime.LocalTime` instead  `java.time.LocalTime`

+++ modules/database/src/commonMain/kotlin/com/mshdabiola/seriesdatabase/model/CourseGradeEntity.kt
@@ -37,10 +37,10 @@ import com.mshdabiola.seriesdatabase.util.Converters
 @TypeConverters(Converters::class)
 data class CourseGradeEntity(
     @ColumnInfo(name = "student_id_fk") // Foreign Key column, Part of Composite PK
-    val studentId: Int, // FK to Student.studentId
+    val studentId: Long, // FK to Student.studentId
 
     @ColumnInfo(name = "course_id_fk") // Foreign Key column, Part of Composite PK
-    val courseId: Int, // FK to Course.courseId
+    val courseId: Long, // FK to Course.courseId
 
     @PrimaryKey // Part of Composite PK, needs @PrimaryKey to be recognized, even in composite key
     @ColumnInfo(name = "academic_year")
